Amber ARM-compatible: opencores.org/project,amber
The Amber processor core is an ARM-compatible 32-bit RISC processor.
The Amber core is fully compatible with the ARM® v2a instruction set
architecture (ISA) and is therefore supported by the GNU toolset. The
Amber project provides a complete embedded system incorporating the
Amber core and a number of peripherals, including a UART, a timer and
an Ethernet MAC.